requirements.txt
fastapi
uvicorn
openai
numpy
tiktoken
pypdf
python-docx
sqlalchemy
faiss-cpu
requests
python-dotenv
python-jose 
